+++ lib/ex_aws/ses.ex
@@ -52,6 +52,43 @@ defmodule ExAws.SES do
   def list_custom_verification_email_templates(opts \\ []) do
     params = build_opts(opts, [:max_results, :next_token])
     request(:list_custom_verification_email_templates, params)
+
+  ## Templates
+  ######################
+
+  @doc """
+  Create an email template.
+  """
+  @type create_template_opt :: {:configuration_set_name, String.t()}
+  @spec create_template(binary, binary, binary, binary, opts :: [create_template_opt]) :: ExAws.Operation.Query.t()
+  def create_template(template_name, subject, html, text, opts \\ []) do
+    template = %{
+      "TemplateName" => template_name,
+      "SubjectPart" => subject
+    }
+    |> put_if_not_nil("HtmlPart", html)
+    |> put_if_not_nil("TextPart", text)
+    |> flatten_attrs("Template")
+
+    params =
+      opts
+      |> build_opts([:configuration_set_name])
+      |> Map.merge(template)
+
+    request(:create_template, params)
+  end
+
+  @doc """
+  Delete an email template.
+  """
+  @spec delete_template(binary) :: ExAws.Operation.Query.t()
+  def delete_template(template_name) do
+    params = %{
+      "TemplateName" => template_name
+    }
+
+    request(:delete_template, params)
+
   end
 
   ## Emails
@@ -64,6 +101,8 @@ defmodule ExAws.SES do
           subject: %{data: binary, charset: binary}
         }
   @type destination :: %{to: [email_address], cc: [email_address], bcc: [email_address]}
+
+  @type bulk_destination :: [%{destination: destination, replacement_template_data: binary}]
 
   @type send_email_opt ::
           {:configuration_set_name, String.t()}
@@ -79,14 +118,6 @@ defmodule ExAws.SES do
   @spec send_email(dst :: destination, msg :: message, src :: binary, opts :: [send_email_opt]) ::
           ExAws.Operation.Query.t()
   def send_email(dst, msg, src, opts \\ []) do
-    dst =
-      Enum.reduce([:to, :bcc, :cc], %{}, fn key, acc ->
-        case Map.fetch(dst, key) do
-          {:ok, val} -> Map.put(acc, :"#{key}_addresses", val)
-          _ -> acc
-        end
-      end)
-
     params =
       opts
       |> build_opts([:configuration_set_name, :return_path, :return_path_arn, :source_arn, :bcc])
@@ -140,23 +171,15 @@ defmodule ExAws.SES do
           opts :: [send_templated_email_opt]
         ) :: ExAws.Operation.Query.t()
   def send_templated_email(dst, src, template, template_data, opts \\ []) do
-    dst =
-      Enum.reduce([:to, :bcc, :cc], %{}, fn key, acc ->
-        case Map.fetch(dst, key) do
-          {:ok, val} -> Map.put(acc, :"#{key}_addresses", val)
-          _ -> acc
-        end
-      end)
-
     params =
       opts
       |> build_opts([:configuration_set_name, :return_path, :return_path_arn, :source_arn, :bcc])
       |> Map.merge(format_member_attribute(:reply_to_addresses, opts[:reply_to]))
       |> Map.merge(format_tags(opts[:tags]))
       |> Map.merge(format_dst(dst))
-      |> Map.put_new("Source", src)
-      |> Map.put_new("Template", template)
-      |> Map.put_new("TemplateData", template_data)
+      |> Map.put("Source", src)
+      |> Map.put("Template", template)
+      |> Map.put("TemplateData", format_template_data(template_data))
 
     request(:send_templated_email, params)
   end
@@ -174,6 +197,34 @@ defmodule ExAws.SES do
       |> Map.put_new("EmailAddress", email_address)
       |> Map.put_new("TemplateName", template_name)
     request(:send_custom_verification_email, params)
+
+  Send a templated email to multiple destinations.
+  """
+  @type send_bulk_templated_email_opt ::
+          {:configuration_set_name, String.t()}
+          | {:return_path, String.t()}
+          | {:return_path_arn, String.t()}
+          | {:source_arn, String.t()}
+          | {:default_template_data, String.t()}
+          | {:tags, %{(String.t() | atom) => String.t()}}
+
+  @spec send_bulk_templated_email(
+          template :: binary,
+          source :: binary,
+          destinations :: bulk_destination,
+          opts :: [send_bulk_templated_email_opt]
+        ) :: ExAws.Operation.Query.t()
+  def send_bulk_templated_email(template, source, destinations, opts \\ []) do
+    params =
+      opts
+      |> build_opts([:configuration_set_name, :return_path, :return_path_arn, :source_arn, :default_template_data])
+      |> Map.merge(format_tags(opts[:tags]))
+      |> Map.merge(format_bulk_destinations(destinations))
+      |> Map.put("DefaultTemplateData", format_template_data(opts[:default_template_data]) )
+      |> Map.put("Source", source)
+      |> Map.put("Template", template)
+
+    request(:send_bulk_templated_email, params)
   end
 
   @doc "Deletes the specified identity (an email address or a domain) from the list of verified identities."
@@ -235,12 +286,45 @@ defmodule ExAws.SES do
     )
   end
 
-  defp format_dst(dst) do
+  defp format_dst(dst, root \\ "destination") do
+    dst = Enum.reduce([:to, :bcc, :cc], %{}, fn key, acc ->
+      case Map.fetch(dst, key) do
+        {:ok, val} -> Map.put(acc, :"#{key}_addresses", val)
+        _ -> acc
+      end
+    end)
+
     dst
     |> Map.to_list()
     |> format_member_attributes([:bcc_addresses, :cc_addresses, :to_addresses])
-    |> flatten_attrs("destination")
+    |> flatten_attrs(root)
   end
+
+  defp format_template_data(nil), do: "{}"
+
+  defp format_template_data(template_data), do: Poison.encode!(template_data)
+
+  defp format_bulk_destinations(destinations) do
+    destinations
+    |> Enum.with_index(1)
+    |> Enum.flat_map(fn
+      {%{destination: destination} = destination_member, index} ->
+        root = "Destinations.member.#{index}"
+
+        destination
+        |> format_dst("#{root}.Destination")
+        |> add_replacement_template_data(destination_member, root)
+        |> Map.to_list()
+    end)
+    |> Map.new()
+  end
+
+  defp add_replacement_template_data(destination, %{replacement_template_data: replacement_template_data}, root) do
+    destination
+    |> Map.put("#{root}.ReplacementTemplateData", format_template_data(replacement_template_data))
+  end
+
+  defp add_replacement_template_data(destination, _, _), do: destination
 
   defp format_tags(nil), do: %{}
 
@@ -310,4 +394,7 @@ defmodule ExAws.SES do
   defp do_flatten_attrs({val, path}) do
     {camelize_key(path), val}
   end
+
+  defp put_if_not_nil(map, _, nil), do: map
+  defp put_if_not_nil(map, key, value), do: map |> Map.put(key, value)
 end