+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v2.1.0
+  - Added support for [CreateTemplate](https://docs.aws.amazon.com/ses/latest/APIReference/API_CreateTemplate.html) action by @themerch
+  - Added support for [DeleteTemplate](hhttps://docs.aws.amazon.com/ses/latest/APIReference/API_DeleteTemplate.html) action by @themerch
+  - Added support for [DeleteTemplate](hhttps://docs.aws.amazon.com/ses/latest/APIReference/API_SendBulkTemplatedEmail.html) action by @themerch
+  -
 v2.0.2
   - Added support for [SendTemplatedEmail](https://docs.aws.amazon.com/ses/latest/APIReference/API_SendTemplatedEmail.html) action with exception for two optional parameters: `TemplateArn` and `ReplyToAddresses.member.N`. by @xfumihiro
   - Fixed wrong key in destination typespec by @kalys